# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "CRUDL-операции по работке с подписками.",
        "title": "subscription-service",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/subscription/create": {
            "post": {
                "description": "Создание новой подписки, subscription_id - primary key. user_id и start_date можно не указывать, на них стоит default. end_date также можно не указывать, в бд допускается null",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "subscription"
                ],
                "summary": "Регистрация новой подписки",
                "parameters": [
                    {
                        "description": "Заполните поля согласно описанию",
                        "name": "subscription",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/subscription-service_internal_entity.SubscriptionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное создание",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Ошибка",
                        "schema": {
                            "$ref": "#/definitions/subscription-service_internal_entity.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка",
                        "schema": {
                            "$ref": "#/definitions/subscription-service_internal_entity.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/subscription/delete": {
            "delete": {
                "description": "Удаление подписки по subscription_id",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "subscription"
                ],
                "summary": "Удаление подписки",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id подписки",
                        "name": "subscription_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное удаление",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка",
                        "schema": {
                            "$ref": "#/definitions/subscription-service_internal_entity.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/subscription/get": {
            "get": {
                "description": "Получение подписки по subscription_id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscription"
                ],
                "summary": "Получение подписки",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id подписки",
                        "name": "subscription_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное получение подписки",
                        "schema": {
                            "$ref": "#/definitions/subscription-service_internal_entity.SubscriptionResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка",
                        "schema": {
                            "$ref": "#/definitions/subscription-service_internal_entity.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/subscription/getList": {
            "get": {
                "description": "Получение всех подписок",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscription"
                ],
                "summary": "Получение всех подписок",
                "responses": {
                    "200": {
                        "description": "Успешное получение подписки",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/subscription-service_internal_entity.SubscriptionResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка",
                        "schema": {
                            "$ref": "#/definitions/subscription-service_internal_entity.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/subscription/update": {
            "put": {
                "description": "Обновление подписки, subscription_id - primary key. service_name, user_id не учитывается. start_date можно не указывать, стоит default. end_date также можно не указывать, в бд допускается null",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "subscription"
                ],
                "summary": "Обновление подписки",
                "parameters": [
                    {
                        "description": "Заполните поля согласно описанию",
                        "name": "subscription",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/subscription-service_internal_entity.SubscriptionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное обновление",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Ошибка",
                        "schema": {
                            "$ref": "#/definitions/subscription-service_internal_entity.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка",
                        "schema": {
                            "$ref": "#/definitions/subscription-service_internal_entity.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "subscription-service_internal_entity.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "subscription-service_internal_entity.SubscriptionRequest": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "service_name": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "subscription_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "subscription-service_internal_entity.SubscriptionResponse": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "service_name": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        }
    }
}